Added storescp option --max-associations.
Added a new command line option --max-associations to storescp that allows
the user to limit the number of network associations handled in parallel
when storescp is running in --fork mode. When the maximum number of child
processes is running, further incoming requests are kept on the listen
backlog until either a child process has ended or the request times out.
On Posix systems, this function uses the SIGCHLD signal to keep track of
child processes. On Windows, a list of child processes together with
RegisterWaitForSingleObject() is used instead.