Fix crash when worker thread reports to the owner pool after destructor of the pool was called.
There is no guarantee that all worker threads have completed their work when when m_runMode is set to SHUTDOWN.
Added an additional condition to the while loop to wait until all workers have completed their tasks and reported to the pool.
After that it's guarantied that all workers are in the m_workerIdle list and can be safely joined.