Fix potential crashes and endless loop.
Fix crash when worker thread reports to the owner pool after destructor
of the pool was called.

There is no guarantee that all worker threads have completed their work
when when m_runMode is set to SHUTDOWN. Added an additional condition to
the while loop to wait until all workers have completed their tasks and
reported to the pool. After that it's guarantied that all workers are in
the m_workerIdle list and can be safely joined.

---

Fixed two cases that make destructor of DcmBaseSCPPool never exit.

Current implementation of DcmBaseSCPPool's destructor waits for the
m_runMode to be set to SHUTDOWN. If m_runMode never set to SHUTDOWN the
destructor will wait forever in the 'while' cycle. Two cases when
m_runMode is not set to SHUTDOWN:

Case 1: The m_runMode won't be set to SHUTDOWN if
DcmBaseSCPPool::listen() was never called.

Fix:

Changed default value of m_runMode from LISTEN to SHUTDOWN. The
m_runMode is set to LISTEN inside of the DcmBaseSCPPool::listen()
function.

Case 2:

Call to initializeNetwork() in the DcmBaseSCPPool::listen() function
might return bad condition. In that case m_runMode won't be set to
SHUTDOWN due to early return.

Fix:

Add a call to the newly added finishListening() function in case of
early return.

Thanks to GitHub user Vovasch for the report and patch.