Fixed bug in handling of odd-length data elements.
When a dataset containing an illegal odd-length attribute with a text VR
was read from file or received over a network connection, then accessing
the value of that attribute with DcmElement::getString() may return a
pointer to a string that was not properly null terminated. Using C string
functions such as strlen() or strcpy() on that string then lead to a read
beyond the end of a string, causing a segmentation fault.

Thanks to Zou Dikai <zoudikai@outlook.com> for the bug report and POC.

This closes DCMTK issue #1184.